refactor and add a test
