+++ src/main.py
@@ -4,9 +4,10 @@ from datetime import datetime, timedelta, timezone
 from typing import List, Optional
 from uuid import UUID, uuid4
 
-import numpy as np
 from fastapi import FastAPI
 from pydantic import BaseModel, Field, validator
+
+from utils import datetime_must_have_timezone, convert_to_camelcase, floor_30_minutes_dt
 
 logger = logging.getLogger(__name__)
 
@@ -30,19 +31,6 @@ app = FastAPI(
 )
 
 thirty_minutes = timedelta(minutes=30)
-
-
-def datetime_must_have_timezone(cls, v: datetime):
-    """Enforce that this variable must have a timezone"""
-    if v.tzinfo is None:
-        raise ValueError(f"{v} must have a timezone")
-    return v
-
-
-def convert_to_camelcase(snake_str: str) -> str:
-    """Converts a given snake_case string into camelCase"""
-    first, *others = snake_str.split("_")
-    return "".join([first.lower(), *map(str.title, others)])
 
 
 class EnhancedBaseModel(BaseModel):
@@ -139,7 +127,7 @@ def _create_dummy_forecast_for_location(location: Location):
 
     # get datetime right now
     now = datetime.now(timezone.utc)
-    now_floor_30 = _floor_30_minutes_dt(dt=now)
+    now_floor_30 = floor_30_minutes_dt(dt=now)
 
     # make list of datetimes that the forecast is for
     datetimes_utc = [now_floor_30 + i * thirty_minutes for i in range(4)]
@@ -245,21 +233,3 @@ async def get_nationally_aggregated_forecasts() -> Forecast:
     return _create_dummy_national_forecast()
 
 
-def _floor_30_minutes_dt(dt):
-    """
-    Floor a datetime by 30 mins.
-
-    For example:
-    2021-01-01 17:01:01 --> 2021-01-01 17:00:00
-    2021-01-01 17:35:01 --> 2021-01-01 17:30:00
-
-    :param dt:
-    :return:
-    """
-    approx = np.floor(dt.minute / 30.0) * 30
-    dt = dt.replace(minute=0)
-    dt = dt.replace(second=0)
-    dt = dt.replace(microsecond=0)
-    dt += timedelta(minutes=approx)
-
-    return dt
